refactor: global.response 와 global.dto에 중복으로 생성되어있는 Rsdata를 하나로 통일
