+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/payment/controller/PaymentController.kt
@@ -3,7 +3,7 @@ package com.NBE_4_5_2.Team5.domain.payment.controller
 import com.NBE_4_5_2.Team5.domain.payment.dto.PaymentDto
 import com.NBE_4_5_2.Team5.domain.payment.dto.PaymentMetaData
 import com.NBE_4_5_2.Team5.domain.payment.service.PaymentService
-import com.NBE_4_5_2.Team5.global.dto.RsData
+import com.NBE_4_5_2.Team5.global.response.RsData
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
