+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/user/logging/LoginLogAspect.kt
@@ -1,7 +1,7 @@
 package com.NBE_4_5_2.Team5.domain.user.logging
 
 import com.NBE_4_5_2.Team5.domain.user.user.controller.UserController.*
-import com.NBE_4_5_2.Team5.global.dto.RsData
+import com.NBE_4_5_2.Team5.global.response.RsData
 import jakarta.servlet.http.HttpServletRequest
 import lombok.RequiredArgsConstructor
 import lombok.extern.slf4j.Slf4j
