+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/aspect/ResponseAspect.kt
@@ -1,6 +1,6 @@
 package com.NBE_4_5_2.Team5.global.aspect
 
-import com.NBE_4_5_2.Team5.global.dto.RsData
+import com.NBE_4_5_2.Team5.global.response.RsData
 import jakarta.servlet.http.HttpServletResponse
 import org.aspectj.lang.ProceedingJoinPoint
 import org.aspectj.lang.annotation.Around
