+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/security/SecurityConfig.kt
@@ -1,7 +1,7 @@
 package com.NBE_4_5_2.Team5.global.security
 
 import com.NBE_4_5_2.Team5.global.dto.Empty
-import com.NBE_4_5_2.Team5.global.dto.RsData
+import com.NBE_4_5_2.Team5.global.response.RsData
 import com.NBE_4_5_2.Team5.global.standard.util.Ut
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.context.annotation.Bean
